Relocate Dragbar to InsertsBar. Cleanup.
