+++ packages/remixer/src/components/index.js
@@ -1,8 +1,8 @@
 export * from './ClearDialog';
-export * from './Dragbar';
 export * from './InsertSlide';
 export * from './InsertTitle';
 export * from './InsertTransition';
+export * from './InsertsBar';
 export * from './LibraryTopbar';
 export * from './RecursiveMenuItem';
 export * from './RemixTopbar';
